+++ decidim-verifications/lib/decidim/verifications.rb
@@ -28,8 +28,8 @@ module Decidim
 
   module Verifications
     include ActiveSupport::Configurable
-    config_accessor :document_types do
-      %w(identification_number,passport)
+    config_accessor :document_type do
+      %w(identification_number passport)
     end
   end
 end