+++ RELEASE_NOTES.md
@@ -56,7 +56,7 @@ decidim_default: &decidim_default
   application_name: <%%= Decidim::Env.new("DECIDIM_APPLICATION_NAME", "My Application Name").to_json %>
   (...)
   verifications:
-    document_types: <%%= Decidim::Env.new("VERIFICATIONS_DOCUMENT_TYPES", %w(identification_number passport)).to_array.to_json %>
+    document_types: <%%= Decidim::Env.new("VERIFICATIONS_DOCUMENT_TYPES", %w(identification_number passport)).to_array %>
 ```
 
 ## 2. Add to your `config/initializers/decidim.rb` the following snippet in the bottom of the file